+++ 05-ScenceManager/BoomerangWeapon.h
@@ -10,7 +10,7 @@ class BoomerangWeapon : public Weapon
 {
 public:
 	int Time;
-	BoomerangWeapon(float nx, float x, float y);
+	BoomerangWeapon(int nx, float x, float y);
 	~BoomerangWeapon();
 	void GetBoundingBox(float& l, float& t, float& r, float& b);
 	void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects = NULL);
